Refactored ContentViewModel file
Moved factories into a dedicated folder and moved CharactersManager into misc folder.
