+++ Rick-and-Morty/Rick And Morty/ViewModels/ContentViewModel.swift
@@ -30,65 +30,13 @@ final class ContentViewModel: ObservableObject {
         var characterViewStates: [CharacterViewState] = []
         
         for character in characters {
-            let characterViewState = characterViewStateFactory.createViewState(from: character)
+            let characterViewState = characterViewStateFactory.createCharacterViewState(from: character)
             characterViewStates.append(characterViewState)
         }
         
         return characterViewStates
     }
 }
-
-
-final class CharactersManager {
-    func fetchCharacters(characterType: CharacterType) -> [Character] {
-        return characterType == .rick ? ricks : morties
-    }
-}
-
-final class CharacterViewStateFactory {
-    func createViewState(from character: Character) -> CharacterViewState {
-        let shortDescription = getShortDescription(for: character)
-        let imagePosition = getImagePosition(for: character)
-        
-        return CharacterViewState(name: character.name, description: character.description, shortDescription: shortDescription, imageName: character.image, imagePosition: imagePosition)
-    }
-    
-    private func getShortDescription(for character: Character) -> String? {
-        if let character = character as? Morty {
-            return character.shortDescription
-        }
-        
-        return nil
-    }
-    
-    private func getImagePosition(for character: Character) -> CharacterImagePosition {
-        if character is Morty {
-            return .left
-        } else {
-            return .right
-        }
-    }
-}
-
-
-
-final class CharactersViewStateFactory {
-    func createCharactersViewState(characterType: CharacterType, characterViewStates: [CharacterViewState]) -> CharactersViewState {
-        let title = getTitle(for: characterType)
-        let iconName = getIconName(for: characterType)
-        
-        return CharactersViewState(title: title, iconName: iconName, characterViewStates: characterViewStates)
-    }
-    
-    private func getTitle(for characterType: CharacterType) -> String {
-        return characterType == .rick ? "Ricks" : "Morties"
-    }
-    
-    private func getIconName(for characterType: CharacterType) -> String {
-        return characterType == .rick ? "rick-icon" : "morty-icon"
-    }
-}
-
 
 struct CharactersViewState: Identifiable {
     let id = UUID()
