+++ Rick-and-Morty/Rick And Morty/misc/CharactersManager.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+final class CharactersManager {
+    func fetchCharacters(characterType: CharacterType) -> [Character] {
+        return characterType == .rick ? ricks : morties
+    }
+}
